Fixed(?) are_lines_intersecting_inf, altered delauanay behavior and it works as intended now
Realized it is wrong to add the faces of the polygon is the input AFL in the constrained case. What needs to happen instead is that the "hull" needs to become its own concept. This is necessary anyway since I am not checking if an edge is part of the convex hull - this is what needs to be added next. When constrained, the hull is provided as the polygon. When unconstrained, the hull must be calculated.

This means that if desired faces are to become part of the triangulation, they can be added to the input AFL themselves.